Add support for whitelist admins
The logic in this support has to a large degree been adapted from the
Honesto token setup which was written for Ethereum/ERC20. It follows
this logic as it introduces the concept of a non-revocable whitelisting
admin which should reduce the risk that the ability to act as a
whitelist admin is not lost for the deployed contract.