Add test of token_metadata_registry endpoint
This test checks that the FA2 contract returns the expected value for
the token_metadata_registry endpoint which is the FA2 contract address.
This endpoint is used for inter-contract communication so it's return
value is a contract call to the caller. That the correct address is
reported in the return value is verified by storing this value in a
new wrapper contract that receives the callback.