Prepare wrapper for handling multiple callback values
This is done by changing the type of storage of the wrapper from `address`
to a record where there is a field for each callback value. For weird
reasons, the deployment process/JS integration required that I added a
dummy-field to the record, since it got converted to the type of the
field if there only was one field. So somewhere in the process, the
integration does something "smart" where it reduces/changes the type of
the record. I think this is something that happens in Truffle.