Add test of balance_of endpoint
This was the only endpoint not tested, so we now have the capability of
testing all endpoints.

Like the token_metadata_registry endpoint, this was tested through a
smart contract that contains a wrapper call for the balance_of endpoint
in the FA2 contract with whitelisting. This wrapper smart contract makes
the call to the FA2 contract, receives the callback, and stores the
callback values in its storage, such that the callback values can be
inspected by fetching the storage and accessing the relevant fields
therein.