Refactor whitelist helper methods
We extract the common functions into a util module. To prevent
passing global state objects around, they simply return the
constructed arguments. This also makes it clearer from just
the code what is going on in the test. We take the
opportunity to perform the `.pkh` access inside the function.