Allow negative KISS balances
This allows the accounts in the KISS contracts to get negative balances
through the signing and registration of tandems. This is a functional
requirement from di Giorgio. The FA1.2 endpoints behave normally and
does not allow negative balances, only the tandem registration does. The
negative balance is implemented as a `debit` field on the account
struct meaning that the code from the fa1.2 partial base implementation
could not be reused like it has in the other FA1.2 implemenations.