Add field in storage for activity log address
This means that we do not have to send the address of the activity log
contract in every call to an endpoint in the KISS FA1.2 contract. This
simplifies the web frontend, but it requires us to to be able to change
this activity log address after the deployment of the contract since
the actity log contract address is only known after deployment of the
contract.

In short: there is mutual dependencies in the two contracts we are
deploying: the KISS FA1.2 contract and the KISS activity log contract.

Maybe this address could be precalculated, so the activity log contract
address would be known before this contract is deployed? I don't yet
know how to do that, though.